+++ NEWS.md
@@ -13,6 +13,8 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.9.0
 Released on CRAN 201X-XX-XX
+* `vcfR2loci()` now has the option `retrun.alleles = FALSE`.
+* `vcfR2genind()` now has the option `retrun.alleles = FALSE`.
 * Error handling code moved into the C++ functions called by read.vcfR so that errors are thrown earlier when reading a VCF. read.vcfR no longer checks that a file is readable first, which solves issues sometimes seen with shared files. (Issue #109, reported and fixed by @NikNakk).
 * `extract.haps()` did not include the parameter `return.alleles = TRUE` in it's call to `extract.gt()` in the haploid branch of the function. This parameter has now been added. This also affects `vcfR2DNAbin()` which calls this function.
 * `vcfR2genlight()` includes the parameter `...` to pass parameters to `adegenet::df2genind()`.